Version v2

+++ include/control/WifiManager.h
@@ -67,7 +67,7 @@ public:
         WiFi.config(INADDR_NONE, INADDR_NONE, INADDR_NONE, INADDR_NONE);
         WiFi.setHostname(s.c_str());
         WiFi.begin(wifiSsid.c_str(), wifiPassword.c_str());
-        long long mil = millis();
+        unsigned long mil = millis();
         while (WiFi.status() != WL_CONNECTED && (millis() - mil > 4000)) {
             delay(500);
             Serial.print(".");
